Extract timer out of world class.
